atoi and atoll fixed

+++ utils/ft_atoi.c
@@ -21,5 +21,5 @@ int ft_atoi(const char* str)
         res = (res * 10) + (str[i] - '0');
         i++;
     }
-    return ((int)(res * sign));
+    return (res * sign);
 }
